Add global auth middleware for api keys
